Don't accept wildcard type bounds.
Wildcards aren't a first-class type according to
the spec, which claims that they "... must appear as a type argument
of a parameterized type.".

Putting them in the bounds of another wildcard causes a position-
invalid `ExistentialTypeTree` to result, since the `whereClauses`
all claim to contain the positions of later wildcards.

Also scouted the grammar of that spec fragment.

Fixes scala/bug#10733.