lub/glb uses correct kind for higher-order type args
When computing the `lub/glb` of HKTs take care to produce
well-kinded types when merging type constructor arguments
that are supplied for a type constructor parameter.

This improves type inference.

A few tests with `-Xstrict-inference` were changed.
Other tests reveal a limitation of higher-kinded existentials.

Fixes scala/bug#8882 and promotes scala/bug#10700 to an error.