Invalidate jar and zip classpath entries reliably
This commit adds a better heuristic to detect changes in classpath
entries. It adds its own definition of "change" in a given file that is
a derivation of what has been agreed in https://github.com/scala/scala/pull/6314.

The implementation is accompanied by test which are in charge of
checking that modifying a CRC or changing the file size make
`FileBasedCache` generate another classpath. As explained in the
comments, if any problem is detected while reading up the CRCs from
every jar entry, the implementation fallbacks on the last modified time.

To check the performance impact of this new detection algorithm there is
a test that naively measures the time to check changes in the classpath
(which happens in every new compile iteration) is kept to a minimum.
Running the test from the sbt shell in a cold JVM process in my Linux
machine measures the duration of the process to be 142ms, a negligible
time span.