Downwards comparisons for implicit search and overloading resolution
This is a backport of the following Dotty change to scalac,

https://github.com/lampepfl/dotty/commit/89540268e6c49fb92b9ca61249e46bb59981bf5a

As in the Dotty implementation the specificity comparison which is used
for overload resolution and implicit selection is now performed as-if
all contravariant positions in top level type constructors were
covariant.

Currently this breaks test/files/run/t2030.scala in exactly the same way
as in Dotty as reported here,

https://github.com/lampepfl/dotty/pull/1246#issuecomment-220710595

and in this PR it has been changed to a neg test. However,
test/files/run/t2030.scala passes when this PR is combined with,

https://github.com/scala/scala/pull/6139

so if/when that is merged it can be switched back to a pos test.

Fixes scala/bug#2509. Fixes scala/bug#7768.