Remove resource hungry tests
Tests that take 30+ seconds to compile and/or execute
really add up to a productivity loss for contributors.

These tests served a purpose to show that the corresponding
changes were correct. But if we want them to serve an ongoing
purpose of guarding against regressions, they need to be
maintained to do so more quicky or be moved into a test suite
that is run less frequently.