Fixes scala/bug#8150. inconsistent Float hashCode
Fixes scala/bug#8150

Before:

```
scala> (1L << 31).## == (1L << 31).toFloat.##
res2: Boolean = false
```

Given that the float value is identical to the long value, hashCode should be the same. As pointed by sjrd, there's a bug in Statics.java that's not catching when Int overflows to `Int.MaxValue` (2147483647), then happens to "equal" with the original `fv` (2147483648) when it gets converted to float.

Note that integers from 2147483584 to 2147483646 will also round up to 2.14748365E9f when converted to `Float`, which then round to `Int.MaxValue` (2147483647). Thus they will return `true` for `==` but hashCode will not match up.