Handle escaping characters in URL -> file
This bug was discovered in
https://github.com/scala/scala/pull/6314#issuecomment-371385148 when
caching classloaders for compiler plugins. The issue discovered is
unrelated to that PR and is rather a bug in how scalac converts URLs to
File in the `AbstractFile.getURL` method.

That method was using `getPath` in `java.net.URL` when the returned path
gives back a file path with escaped characters. When that incorrect file
path is passed through the constructor of `java.io.File`, the
construction succeeds but the underlying file is a different one (and
doesn't exist).

The fix to this bug is to use the safe `toURI()` method, which is
correctly used in other parts of the scalac classpath infrastructure
(like in the `asURLs` method in `DirectoryClasspath` and
`ZipArchiveFileLookup`). The `toURI` method returns the file path with
all the special characters unescaped.

Fixes https://github.com/scala/bug/issues/10764.