Put back one-argument require for the convenience of downstream users.
I was inspired by paulp's comment years ago that we can remove the
one-argument forms of `assert` and `require` by shadowing them with
two-argument versions in `Global` (now `SymbolTable`). I tried
carrying out that hope, and was reminded by the community build
that people write compiler plugins against the compiler "API", so
that just won't fly.

I'm advised that it's kinder to put them back than to make plugin
writers scramble to change their code in order to release for 2.12.5,
so let's do that.

I'd also like us to consider deprecating them in a future release,
since raw assertions make guessing what went wrong harder.
Or not; it's not my call to make.