Singleton typed args/results don't interfere with tailcalls
At one point the literal types PR modified the semantics of asInstanceOf
to check for equality of the target value with the value embedded in the
singleton type. This interacted badly with mkAttributedCastHack in
Tailcalls,

  https://github.com/dotta/scala/commit/c2534bf61b61b76cebaad09ba303234193709b60

because the inserted casts could end up referring to a non-existent
Symbol which would be needed to implement the equality check during
Erasure.

The fix at the time was to make the use of a singleton type as a method
argument or result type disqualify the method from having the tailcall
optimization applied. A later decision by the SIP committee decided for
unrelated reasons that an asInstanceOf test should not perform an
equality check, so this disqualification became unnecessary. It was not
removed, however.

This came to light because singleton types are used as method argument
and result types in methods which are expected to have the tail call
optimizatino applied, resulting in,

  https://github.com/scala/scala-dev/issues/467

We fix that here by doing what should have been done earlier: removing
the disqualification.