Handle rangepos of by-name right-associative applications properly.
And, as a bonus, those with infix type args.

The by-name case is a simple tweak to the SIP-34 translation:
the pos of the newly-typed `Apply` needs to encompass the inlined
rhs of the synthetic `ValDef`. `wrappingPos` was made for this.

For right-associative applications with explicit type args, the
situation gets pretty hairy in that we have something akin to:

```
argument operator_: [TYPEARG] qualifier
                    ^-Ident-^
         ^-----------Select-----------^
         ^---------TypeApply----------^
^---------------Apply-----------------^
```

that parses to

```
Apply(
  TypeApply(
    Select(qualifier, operator_:),
    Ident(TYPEARG) :: Nil
  ),
  argument :: Nil
)
```

The problem is that the type argument falls within the `Select`,
so both the `fun` and `targs` of the `TypeApply` contain that
position. `makeTransparent` was made for this. `makeTransparentIf`
was made for concision.

Fixes scala/bug#10751.