Keep sbt's log level at info on travis / jenkins
Setting `--warn` hides the `downloading ... scala-compiler-2.12.4.jar`
messages (which I was looking for to ensure clearing the caches has
the intended effect).

Compared to the current state, it doesn't add a significant amount of
noise on travis. Note that the travis log has folded sections, so
it's easy to look at individual parts.

There are still `--warn` arguments in the pr validation and windows
scripts that I didn't touch here.