Treat JAVA_CONST names as snake for completion
The previous behavior is for every uppercase
letter to represent a camel hump, so "jon"
and "a_o" would complete to "JAVA_CONST".

This commit splits the identifier on underscore
if the user is not asking for underscore,
such as "_local", and if the candidate name looks
like an old-style Java constant, uppercase
with underscores.