Fix return value of sbtResolve
`sbtResolve` is used in an `if` test. The last statement of the function
makes the return value, so the `travis_fold_end` made it univerally
`true`.

For reference, the `errexit` mode (`set -e`, set in `common`) is
ignored in the test position of an `if` statement (otherwise the
build would have been aborted), see `man bash`.