use ScalaCheck built for M3
I assume it was inadvertent that we were still using an M1 build
of ScalaCheck.  It wasn't causing any problems we know of,
but we should still be on the newer one.

also it's not clear why the dependency was hardcoded instead of
going through `scalaDep`, so I changed that to be consistent
with our other Scala-based dependencies