Build a new starr first
We need 3 build stages, locker+quick is not enough, because of scaladoc
and xml. Say we build scala `vLOCKER` with `vSTARR`, then the
compilation classpath when compiling scaladoc contains `xml_vSTARR`. The
`compiler.pom` will depend on `xml_vSTARR`.

When we build quick, the `doc` command will run `scaladoc-vLOCKER` with
`xml_vSTARR` on the runtime classpath. This is broken because
`library-vLOCKER` might not provide what `xml_vSTARR` expects. In the
the case of the new collections, `scala.collection.SeqLike` is missing.

Note that this issue shows that we need to use the same version number
for quick and locker, otherwise the quick `compiler.pom` will depend on
the wrong version of xml.