[nomerge] Streamline the bootstrap script
Ensure that modules are built in each stage (if they need building),
including even starr. This ensures that the locker scaladoc when
building quick runs with a freshly built xml on the classpath, which
can be necessary for binary compatibility.

Run scaladoc only in the quick stage (not starr or locker), for scala
and the modules.

Don't use `clone --reference` to build starr, doesn't work with shallow
clones (on travis).

Set `STARR_VER` to the full `SCALA_VER` + a suffix, so that it gets
the same treatment in sbt with respect to cross-versioning (binary
vs full).

The quick stage either publishes to scala-integration or sonatype
(for releases).

[nomerge] because this commit is in the 2.13.x-new-collections branch.