Fix up embarrassing typo.
I put this in in cdf74190c442ff60dc6b4ed7c7567fb58448a90e, right
before the PR got merged, and moved the condition that used to be
in the `else` block into a `devWarning`, as I was pretty sure we'd
never get here at or later than erasure (otherwise, there's a chance
we'd let a `Constant(<unerased type>)` get to the backend. I wasn't
willing to wager on it for an assertion, though.

In retrospect it would have been better to poke around with `-Xdev`
on first, so I would notice the reversed condition. (I build with
the flag at work so someone sees the warnings, is how I noticed.)