Support cancellation by checking Thread.interrupted
A common means of cancelling a task is to shutdown the thread
pool executing it. That's what SBT's CTRL-C handler does, for example.

Typically, thread pools call `Thread.interrupt()` to cooperatively
stop the workload. We need to do our part by checking `interrupted()`
from time to time, and translating this into an exception that will
stop compilation.