Honour Thread.interrupt detected by NIO operations
NIO file operations check for thread interruption and throw an
exception. If we see one of these, all subsequent operations will
fail in the same way, so we should fail fast.

Since the change to use NIO for file writing, and prior to this change,
hitting CTRL-C during `sbt compile` would spew out a stack trace
for each classfile.