Update test/files/speclib/instrumented.jar
The file contains patched versions of ScalaRunTime and BoxesRunTime.
The jar can be re-generated by the mkinstrumented.sh, then needs to
be uploaded to bintray.com/typesafe/scala-sha-bootstrap.

Updated the patches that the script applies to use the unified format,
which provides some context, and is less likely to break in the future.

Restored the `arrayUpdateCount` counter. It was removed in a19babc952,
but the specialized tests actually use it. Note that no instrumented.jar
was ever published to bintray with the changes in a19babc952.

partest --specialized passes (also --res, --scalap, --instrumented)