Ensure stability in `Typer.packedType`
Method `packedType` does an existential over local symbols. The locals
are collected in a Set, which is converted `toList`. This is not stable,
but didn't show up in the old collections because of the Set1/Set2
special cases. In the ChampHashSet it started showing for small sets.

This makes `neg/variances-refinement.scala` stable again.