Drop flags that change language semantics
To reduce the many possible combination of these flags,
the only supported mechanism that affects the outcome of
type checking is `-Xsource`, which instructs the compiler
to mimic the given version as much as possible. We only
support this for one prior and one subsequent version
(e.g., 2.13 recognizes 2.12 and 2.14 source levels).

The following flags have been dropped:
  - `-Xfull-lubs`
  - `-Xstrict-inference`
  - `-Yinfer-argument-types`
  - `-Yno-adapted-args`
  - `-Yoverride-objects`
  - `-Yoverride-vars`
  - `-Yvirtpatmat`

We're exceptionally dropping a few obscure `-X` flags
without prior deprecation. `-Y` flags are always at our whim.