Move the patmat phase after superaccessors
In dotty, pickling trees into Tasty is done after the super accessors
phase but before the pattern matching phase. But in scalac, patmat comes
before superaccessors. This commit simply swaps these phases around,
thus making it possible to write a compiler plugin for scalac that
produces a Tasty file from the output of the superaccessors phase.