Drop old @version refs in docs
A lot of the @version entries in the scaladoc comments for the scala
library are no longer maintained.  There was probably a goal to keep
them updated over time.  Since its bundled with the compiler, the
scala library version is based on the compiler.  That's what
determines scala library compatability.

Since the version field is not maintained, the values just look like
unnecessary cruft when the library api docs are published on the web.

The @since version number seems like it's worth preserving.  I've
changed the @version to a @since entry when it was appropriate.