Avoid cyclic error in runtime reflection w. ref to inner class in parent
This brings it into line with the regular compiler. The mechanics are
a bit different: ClassFileParser has an internal map of declared inner
classes of the currently parsed outer class, and any class references
look at this first before looking up the owner's info.

That was a too big a change to contemplate for runtime reflection,
so I've opted instead to assign a parent-free info to the class and
module first, use the existing code to enter the inner classes
(changing it to use rawInfo to avoid the cyclic error detection
in info), and then to add the accumulated member classes in the
real ClassInfo types created below.