Prefer internal.Reporter
Complete the reporting refactor to use internal.Reporter.

The partest blocker was fixed in v1.0.14. sbt and IDE
are not validated, so reporters.Reporter is not removed.

There are internal dependencies on cancel and comment
functionality, which is resolved here by typecasing.
Both features need to be implemented otherwise.

ConsoleReporter is still a reporter.Reporter.