Make Array.copyAs Scala.js-friendly.
When copying between object arrays, `Array.copyAs` needs to produce
a `Class[Array[destClass]]`. The previous version used an
incantation with `Class.forName` to do so, but that is not
supported in Scala.js. It uselessly tested for cases where `c` is
a primitive type, but that never happens because `copyAs` has a
different code path when `destClass.isPrimitive`.

This commit uses `Array.newInstance(destClass, 0).getClass`
instead, which is supported by Scala.js.