Move range position validation behind a flag.
The motivation is that the validation step isn't fast, and takes
up a good chunk of the "rangepos penalty" time difference. Moreover,
Alex Average User can't do much about a fatal rangepos error other
than twiddle around their source until it goes away, so it's likely
to bother end users less like this.

This is a backport to 2.12.x, since position validation is changing
for performance, and we evidently want to be cautious about adding
new breakages.

References scala/scala-dev#472.