Outer parameters have the ACC_SYNTHETIC flag in bytecode.
Apparently the generic signature for constructors is not
expected to mention the outer accessor, but the descriptor
obviously must. This discrepancy must be handled by the
Java reflection method `Parameter#getParameterizedType`,
which knows to ignore synthetic (or "mandated") method
parameters that it sees in the descriptor while parsing
the signature.

This relies heavily on the `MethodParameters` classfile
attribute, and experimentation shows that stripping that
information from the classfile causes `getParameterizedType`
to report only the erased types that it sees in the
descriptor.

Javac, with `-parameters`, emits the outer accessor with
the `ACC_MANDATED` flag, which we don't emit (and doesn't
appear to be a public API yet). However, it interprets
`ACC_SYNTHETIC` in the same way, and we do emit that (now).

This should be a one-liner, but GenBCode reads the parameter
symbols off the `DefDef`, not the method's symbol's info.
This shouldn't matter, but I did notice that we make another,
*different* symbol for the parameter to use in the method's
info. (It's also got a different name: `arg$outer` rather than
`$outer`.) To be safe, I marked them both `ARTIFACT`.

Fixes scala/bug#10880.