Fixes Issue 493, to avoid unnecessary allocation.
This commit avoids unnecessary memory use noticed
in Issue 493 of https://github.com/scala/scala-dev.

The line: `refinementParents :::= currentBaseClass.parentSymbols`
in each iteration pre-prends `currentBaseClass.parentSymbols` to the
previous `refinementParens`. Thus, at the end of the loop,
the length of `refinementParens` is the sum of the `parentSymbols`
lists obtained from each symbol in `initBaseClasses` which is a
refinement classes. That creates as many cons (`::`) objects.
Moreover, since `parentSymbols` is not a `val` but a `def`, it creates
a list of length `m`, copies it (to prepend it) and the throws it away.

To prevent these allocations, we replace the flattened
`refinedParents` list by a `refinedClasses` list, which
stores the classes whose parents we have not yet looked into.
We just use the `exists` methods of the List class to look for a
refinedClass with at least one parent is the currentBaseClass.

(cherry picked from commit 0db7dd4e251ec7be75f6f8178977faae9c8274e9)