Runner reads scalac args from file.
Just like it does for javac.

Som tried to do this a few PRs ago to find that it doesn't work,
and I realized that I hadn't yet realized that it should work.

Now it does.

Also, I found paulp's old `rm-orphan-checkfile` script, and ran
it on a whim. Turns out it still does things.