More precise bounds for *SetOps traits
We now always define them with bounds like this:

    trait FooSetOps[
          A,
          +CC[X] <: FooSetOps[X, CC, _] with FooSet[X],
          +C <: FooSetOps[A, CC, C] with CC[A]
    ]

This is the most precise bounds we can give to this type. Most of the
time this precision is not needed but this makes it easier to add new
operations without breaking compatibility.

The same scheme could also be used for other *Ops traits but I haven't
tried it.