Make Set covariant, create InvariantSetOps
As a first step, we make Set and immutable.Set covariant, but keep their
operations as-is, annotated with @uV to typecheck. We also create
InvariantSetOps and immutable.InvariantSetOps that we will use in
subsequent commits to hold the operations that only make sense on
invariant subclasses of Set. Note that currently, the only non-mutable
Sets that must be invariant are the SortedSets, so we could fusion
InvariantSetOps with SortedSetOps to reduce the API at the cost of some
flexibility.