Adapt apply, contains and +/incl for covariant Set
`contains` now takes an extra type parameter to pass variance checks. To
avoid boxing we overload `contains` in BitSet. The implementation of
`contains` in GenKeySet and TreeSet are currently unsafe and hard to
make safe due to erasure, this will have to be resolved before this PR
can be merged.

`+`/`incl` are changed similarly. Additionally, overloaded versions
with their old type signatures are added to InvariantSetOps. This is
crucial for all the SortedSets since adding an element of the same type
should preserve the sorting.